#143 Split Genre.kt and move parseGenre/joinGenres to top-level functions
Split the ~1200-line Genre.kt into three files in
net.transgressoft.commons.music.audio:
- Genre.kt: sealed class shell + nested Custom only
- GenreDefinitions.kt: the predefined genre data objects (now top-level)
- GenreExtensions.kt: top-level parseGenre/joinGenres + lazy name index

Migrate all call sites off the removed Genre.parseGenre/Genre.joinGenres
companion API to the top-level functions. Parsing/joining behavior is
unchanged (case-insensitive match, Custom fallback, comma split, sorted join).

Source-breaking: genre constants are now top-level (Rock, not Genre.Rock) and
the companion @JvmStatic helpers no longer exist; Java callers use
GenreExtensionsKt.parseGenre(...). Adds GenreExtensionsTest.
