+++ music-commons-api/src/main/kotlin/net/transgressoft/commons/music/audio/GenreExtensions.kt
@@ -1,0 +1,426 @@
+/******************************************************************************
+ * Copyright (C) 2025  Octavio Calleya Garcia                                 *
+ *                                                                            *
+ * This program is free software: you can redistribute it and/or modify       *
+ * it under the terms of the GNU General Public License as published by       *
+ * the Free Software Foundation, either version 3 of the License, or          *
+ * (at your option) any later version.                                        *
+ *                                                                            *
+ * This program is distributed in the hope that it will be useful,            *
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of             *
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the              *
+ * GNU General Public License for more details.                               *
+ *                                                                            *
+ * You should have received a copy of the GNU General Public License          *
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.     *
+ ******************************************************************************/
+
+package net.transgressoft.commons.music.audio
+
+// Hardcoded map avoids kotlin-reflect dependency on Genre::class.sealedSubclasses.
+// Lazy initialization is required because the data object instances are not yet initialized
+// when this top-level property's initializer would otherwise run during class loading.
+private val BY_NAME: Map<String, Genre> by lazy {
+    buildMap {
+        put(Abstract.name.lowercase(), Abstract)
+        put(AbstractHipHop.name.lowercase(), AbstractHipHop)
+        put(AbstractRock.name.lowercase(), AbstractRock)
+        put(Acapella.name.lowercase(), Acapella)
+        put(AcidBreakbeat.name.lowercase(), AcidBreakbeat)
+        put(AcidHouse.name.lowercase(), AcidHouse)
+        put(AcidJazz.name.lowercase(), AcidJazz)
+        put(AcidRock.name.lowercase(), AcidRock)
+        put(Acoustic.name.lowercase(), Acoustic)
+        put(Alternative.name.lowercase(), Alternative)
+        put(AlternativeCountry.name.lowercase(), AlternativeCountry)
+        put(AlternativeDance.name.lowercase(), AlternativeDance)
+        put(AlternativeMetal.name.lowercase(), AlternativeMetal)
+        put(AlternativeRock.name.lowercase(), AlternativeRock)
+        put(Ambient.name.lowercase(), Ambient)
+        put(Americana.name.lowercase(), Americana)
+        put(Anime.name.lowercase(), Anime)
+        put(ArtRock.name.lowercase(), ArtRock)
+        put(Avantgarde.name.lowercase(), Avantgarde)
+        put(AvantgardeJazz.name.lowercase(), AvantgardeJazz)
+        put(AvantgardeMetal.name.lowercase(), AvantgardeMetal)
+        put(Bachata.name.lowercase(), Bachata)
+        put(Baroque.name.lowercase(), Baroque)
+        put(BassMusic.name.lowercase(), BassMusic)
+        put(Beatbox.name.lowercase(), Beatbox)
+        put(Bebop.name.lowercase(), Bebop)
+        put(BigBand.name.lowercase(), BigBand)
+        put(BigBeat.name.lowercase(), BigBeat)
+        put(Bitcore.name.lowercase(), Bitcore)
+        put(Bitpop.name.lowercase(), Bitpop)
+        put(BlackMetal.name.lowercase(), BlackMetal)
+        put(Bluegrass.name.lowercase(), Bluegrass)
+        put(Blues.name.lowercase(), Blues)
+        put(BluesRock.name.lowercase(), BluesRock)
+        put(Boogie.name.lowercase(), Boogie)
+        put(BoogieWoogie.name.lowercase(), BoogieWoogie)
+        put(BossaNova.name.lowercase(), BossaNova)
+        put(BrassBand.name.lowercase(), BrassBand)
+        put(Breakbeat.name.lowercase(), Breakbeat)
+        put(Breakcore.name.lowercase(), Breakcore)
+        put(Britpop.name.lowercase(), Britpop)
+        put(BrokenBeat.name.lowercase(), BrokenBeat)
+        put(BumbaMeuBoi.name.lowercase(), BumbaMeuBoi)
+        put(Celtic.name.lowercase(), Celtic)
+        put(CelticFusion.name.lowercase(), CelticFusion)
+        put(CelticMetal.name.lowercase(), CelticMetal)
+        put(CelticPunk.name.lowercase(), CelticPunk)
+        put(CelticReggae.name.lowercase(), CelticReggae)
+        put(CelticRock.name.lowercase(), CelticRock)
+        put(Chamber.name.lowercase(), Chamber)
+        put(ChamberJazz.name.lowercase(), ChamberJazz)
+        put(ChamberMusic.name.lowercase(), ChamberMusic)
+        put(ChamberPop.name.lowercase(), ChamberPop)
+        put(Chanson.name.lowercase(), Chanson)
+        put(Chant.name.lowercase(), Chant)
+        put(ChaChaCha.name.lowercase(), ChaChaCha)
+        put(Chicha.name.lowercase(), Chicha)
+        put(Children.name.lowercase(), Children)
+        put(Chillout.name.lowercase(), Chillout)
+        put(Chillwave.name.lowercase(), Chillwave)
+        put(Chiptune.name.lowercase(), Chiptune)
+        put(Christian.name.lowercase(), Christian)
+        put(Christmas.name.lowercase(), Christmas)
+        put(CityPop.name.lowercase(), CityPop)
+        put(Classic.name.lowercase(), Classic)
+        put(Classical.name.lowercase(), Classical)
+        put(ClassicPop.name.lowercase(), ClassicPop)
+        put(ClassicRock.name.lowercase(), ClassicRock)
+        put(Club.name.lowercase(), Club)
+        put(Comedy.name.lowercase(), Comedy)
+        put(ContemporaryBlues.name.lowercase(), ContemporaryBlues)
+        put(ContemporaryClassical.name.lowercase(), ContemporaryClassical)
+        put(ContemporaryJazz.name.lowercase(), ContemporaryJazz)
+        put(CoolJazz.name.lowercase(), CoolJazz)
+        put(Country.name.lowercase(), Country)
+        put(CountryPop.name.lowercase(), CountryPop)
+        put(CountryRock.name.lowercase(), CountryRock)
+        put(Crossover.name.lowercase(), Crossover)
+        put(Crunk.name.lowercase(), Crunk)
+        put(Crunkcore.name.lowercase(), Crunkcore)
+        put(CrustPunk.name.lowercase(), CrustPunk)
+        put(CGothic.name.lowercase(), CGothic)
+        put(CHipHop.name.lowercase(), CHipHop)
+        put(CJazz.name.lowercase(), CJazz)
+        put(CPop.name.lowercase(), CPop)
+        put(CRock.name.lowercase(), CRock)
+        put(CSka.name.lowercase(), CSka)
+        put(CTrance.name.lowercase(), CTrance)
+        put(Dance.name.lowercase(), Dance)
+        put(Dancehall.name.lowercase(), Dancehall)
+        put(DarkAmbient.name.lowercase(), DarkAmbient)
+        put(DarkCore.name.lowercase(), DarkCore)
+        put(DarkPop.name.lowercase(), DarkPop)
+        put(DarkStep.name.lowercase(), DarkStep)
+        put(DarkWave.name.lowercase(), DarkWave)
+        put(Deathcore.name.lowercase(), Deathcore)
+        put(Deathgrind.name.lowercase(), Deathgrind)
+        put(DeathMetal.name.lowercase(), DeathMetal)
+        put(DeepFunk.name.lowercase(), DeepFunk)
+        put(DeepHouse.name.lowercase(), DeepHouse)
+        put(DeepSoul.name.lowercase(), DeepSoul)
+        put(Disco.name.lowercase(), Disco)
+        put(DiscoHouse.name.lowercase(), DiscoHouse)
+        put(DixielandJazz.name.lowercase(), DixielandJazz)
+        put(Doomcore.name.lowercase(), Doomcore)
+        put(DoomMetal.name.lowercase(), DoomMetal)
+        put(DooWop.name.lowercase(), DooWop)
+        put(Downtempo.name.lowercase(), Downtempo)
+        put(DreamPop.name.lowercase(), DreamPop)
+        put(Drone.name.lowercase(), Drone)
+        put(DroneMetal.name.lowercase(), DroneMetal)
+        put(DrumAndBass.name.lowercase(), DrumAndBass)
+        put(Dub.name.lowercase(), Dub)
+        put(Dubstep.name.lowercase(), Dubstep)
+        put(DubHouse.name.lowercase(), DubHouse)
+        put(Dubtronica.name.lowercase(), Dubtronica)
+        put(Ebm.name.lowercase(), Ebm)
+        put(Edm.name.lowercase(), Edm)
+        put(Electro.name.lowercase(), Electro)
+        put(Electronic.name.lowercase(), Electronic)
+        put(Electronica.name.lowercase(), Electronica)
+        put(ElectronicDance.name.lowercase(), ElectronicDance)
+        put(Electropop.name.lowercase(), Electropop)
+        put(ElectroSwing.name.lowercase(), ElectroSwing)
+        put(ElectroWave.name.lowercase(), ElectroWave)
+        put(Emo.name.lowercase(), Emo)
+        put(EmoPop.name.lowercase(), EmoPop)
+        put(EmoRap.name.lowercase(), EmoRap)
+        put(Enka.name.lowercase(), Enka)
+        put(Ethnic.name.lowercase(), Ethnic)
+        put(EuroDance.name.lowercase(), EuroDance)
+        put(EuroDisco.name.lowercase(), EuroDisco)
+        put(EuroHouse.name.lowercase(), EuroHouse)
+        put(EuroPop.name.lowercase(), EuroPop)
+        put(EuroTrance.name.lowercase(), EuroTrance)
+        put(Experimental.name.lowercase(), Experimental)
+        put(ExperimentalNoise.name.lowercase(), ExperimentalNoise)
+        put(ExperimentalPop.name.lowercase(), ExperimentalPop)
+        put(ExperimentalRock.name.lowercase(), ExperimentalRock)
+        put(FemaleVocalist.name.lowercase(), FemaleVocalist)
+        put(Fingerstyle.name.lowercase(), Fingerstyle)
+        put(Flamenco.name.lowercase(), Flamenco)
+        put(Folk.name.lowercase(), Folk)
+        put(FolkHop.name.lowercase(), FolkHop)
+        put(FolkMetal.name.lowercase(), FolkMetal)
+        put(FolkPop.name.lowercase(), FolkPop)
+        put(FolkPunk.name.lowercase(), FolkPunk)
+        put(FolkRock.name.lowercase(), FolkRock)
+        put(Folktronica.name.lowercase(), Folktronica)
+        put(FreakFolk.name.lowercase(), FreakFolk)
+        put(Funk.name.lowercase(), Funk)
+        put(Fusion.name.lowercase(), Fusion)
+        put(FusionJazz.name.lowercase(), FusionJazz)
+        put(FutureBass.name.lowercase(), FutureBass)
+        put(FutureJazz.name.lowercase(), FutureJazz)
+        put(Garage.name.lowercase(), Garage)
+        put(GarageRock.name.lowercase(), GarageRock)
+        put(GlamMetal.name.lowercase(), GlamMetal)
+        put(GlamPunk.name.lowercase(), GlamPunk)
+        put(GlamRock.name.lowercase(), GlamRock)
+        put(Glitch.name.lowercase(), Glitch)
+        put(GlitchHop.name.lowercase(), GlitchHop)
+        put(GloFi.name.lowercase(), GloFi)
+        put(Goa.name.lowercase(), Goa)
+        put(Gospel.name.lowercase(), Gospel)
+        put(Gothic.name.lowercase(), Gothic)
+        put(GothicMetal.name.lowercase(), GothicMetal)
+        put(GothicRock.name.lowercase(), GothicRock)
+        put(Grime.name.lowercase(), Grime)
+        put(Grindcore.name.lowercase(), Grindcore)
+        put(Grunge.name.lowercase(), Grunge)
+        put(GypsyJazz.name.lowercase(), GypsyJazz)
+        put(GypsyPunk.name.lowercase(), GypsyPunk)
+        put(GFunk.name.lowercase(), GFunk)
+        put(Hardcore.name.lowercase(), Hardcore)
+        put(Hardstyle.name.lowercase(), Hardstyle)
+        put(HardBop.name.lowercase(), HardBop)
+        put(HardHouse.name.lowercase(), HardHouse)
+        put(HardRock.name.lowercase(), HardRock)
+        put(HardTrance.name.lowercase(), HardTrance)
+        put(HeavyMetal.name.lowercase(), HeavyMetal)
+        put(HipHop.name.lowercase(), HipHop)
+        put(Horrorcore.name.lowercase(), Horrorcore)
+        put(HorrorPunk.name.lowercase(), HorrorPunk)
+        put(House.name.lowercase(), House)
+        put(Idm.name.lowercase(), Idm)
+        put(Illbient.name.lowercase(), Illbient)
+        put(Impressionist.name.lowercase(), Impressionist)
+        put(Incidental.name.lowercase(), Incidental)
+        put(Indie.name.lowercase(), Indie)
+        put(IndieFolk.name.lowercase(), IndieFolk)
+        put(IndiePop.name.lowercase(), IndiePop)
+        put(IndieRock.name.lowercase(), IndieRock)
+        put(Indietronic.name.lowercase(), Indietronic)
+        put(Industrial.name.lowercase(), Industrial)
+        put(IndustrialMetal.name.lowercase(), IndustrialMetal)
+        put(IndustrialRock.name.lowercase(), IndustrialRock)
+        put(Instrumental.name.lowercase(), Instrumental)
+        put(InstrumentalHipHop.name.lowercase(), InstrumentalHipHop)
+        put(InstrumentalRock.name.lowercase(), InstrumentalRock)
+        put(Jazz.name.lowercase(), Jazz)
+        put(JazzFusion.name.lowercase(), JazzFusion)
+        put(JazzHop.name.lowercase(), JazzHop)
+        put(Jumpstyle.name.lowercase(), Jumpstyle)
+        put(Jungle.name.lowercase(), Jungle)
+        put(JFusion.name.lowercase(), JFusion)
+        put(JGothic.name.lowercase(), JGothic)
+        put(JHipHop.name.lowercase(), JHipHop)
+        put(JJazz.name.lowercase(), JJazz)
+        put(JPop.name.lowercase(), JPop)
+        put(JRock.name.lowercase(), JRock)
+        put(JSka.name.lowercase(), JSka)
+        put(JTrance.name.lowercase(), JTrance)
+        put(Krautrock.name.lowercase(), Krautrock)
+        put(KGothic.name.lowercase(), KGothic)
+        put(KHipHop.name.lowercase(), KHipHop)
+        put(KJazz.name.lowercase(), KJazz)
+        put(KPop.name.lowercase(), KPop)
+        put(KRock.name.lowercase(), KRock)
+        put(KSka.name.lowercase(), KSka)
+        put(KTrance.name.lowercase(), KTrance)
+        put(Latin.name.lowercase(), Latin)
+        put(LatinJazz.name.lowercase(), LatinJazz)
+        put(LatinPop.name.lowercase(), LatinPop)
+        put(LatinRock.name.lowercase(), LatinRock)
+        put(LiquidFunk.name.lowercase(), LiquidFunk)
+        put(LoFi.name.lowercase(), LoFi)
+        put(Lounge.name.lowercase(), Lounge)
+        put(Mambo.name.lowercase(), Mambo)
+        put(Marimba.name.lowercase(), Marimba)
+        put(Mathcore.name.lowercase(), Mathcore)
+        put(MathRock.name.lowercase(), MathRock)
+        put(Medieval.name.lowercase(), Medieval)
+        put(MedievalMetal.name.lowercase(), MedievalMetal)
+        put(MedievalRock.name.lowercase(), MedievalRock)
+        put(MelodicMetal.name.lowercase(), MelodicMetal)
+        put(MelodicMetalcore.name.lowercase(), MelodicMetalcore)
+        put(MelodicTrance.name.lowercase(), MelodicTrance)
+        put(Merengue.name.lowercase(), Merengue)
+        put(Metal.name.lowercase(), Metal)
+        put(Metalcore.name.lowercase(), Metalcore)
+        put(Microhouse.name.lowercase(), Microhouse)
+        put(Minimal.name.lowercase(), Minimal)
+        put(ModalJazz.name.lowercase(), ModalJazz)
+        put(ModernClassical.name.lowercase(), ModernClassical)
+        put(ModernRock.name.lowercase(), ModernRock)
+        put(Motown.name.lowercase(), Motown)
+        put(Mpb.name.lowercase(), Mpb)
+        put(Musical.name.lowercase(), Musical)
+        put(MushroomJazz.name.lowercase(), MushroomJazz)
+        put(Ndw.name.lowercase(), Ndw)
+        put(NeoClassical.name.lowercase(), NeoClassical)
+        put(NeoFolk.name.lowercase(), NeoFolk)
+        put(NeoMedieval.name.lowercase(), NeoMedieval)
+        put(NeoProgressive.name.lowercase(), NeoProgressive)
+        put(NeoPsychedelic.name.lowercase(), NeoPsychedelic)
+        put(NeoSoul.name.lowercase(), NeoSoul)
+        put(Nerdcore.name.lowercase(), Nerdcore)
+        put(NewAge.name.lowercase(), NewAge)
+        put(NewWave.name.lowercase(), NewWave)
+        put(Newgrass.name.lowercase(), Newgrass)
+        put(Noise.name.lowercase(), Noise)
+        put(NoisePop.name.lowercase(), NoisePop)
+        put(NoiseRock.name.lowercase(), NoiseRock)
+        put(NuDisco.name.lowercase(), NuDisco)
+        put(NuJazz.name.lowercase(), NuJazz)
+        put(NuMetal.name.lowercase(), NuMetal)
+        put(NuSoul.name.lowercase(), NuSoul)
+        put(Nwa.name.lowercase(), Nwa)
+        put(Nwobhm.name.lowercase(), Nwobhm)
+        put(Nyhc.name.lowercase(), Nyhc)
+        put(Oldies.name.lowercase(), Oldies)
+        put(Opera.name.lowercase(), Opera)
+        put(Orchestral.name.lowercase(), Orchestral)
+        put(Oriental.name.lowercase(), Oriental)
+        put(OutlawCountry.name.lowercase(), OutlawCountry)
+        put(Pop.name.lowercase(), Pop)
+        put(PopPunk.name.lowercase(), PopPunk)
+        put(PopRock.name.lowercase(), PopRock)
+        put(PostBritpop.name.lowercase(), PostBritpop)
+        put(PostDisco.name.lowercase(), PostDisco)
+        put(PostGrunge.name.lowercase(), PostGrunge)
+        put(PostHardcore.name.lowercase(), PostHardcore)
+        put(PostIndustrial.name.lowercase(), PostIndustrial)
+        put(PostMetal.name.lowercase(), PostMetal)
+        put(PostPunk.name.lowercase(), PostPunk)
+        put(PostRock.name.lowercase(), PostRock)
+        put(PowerMetal.name.lowercase(), PowerMetal)
+        put(PowerNoise.name.lowercase(), PowerNoise)
+        put(PowerPop.name.lowercase(), PowerPop)
+        put(ProgressiveBluegrass.name.lowercase(), ProgressiveBluegrass)
+        put(ProgressiveCountry.name.lowercase(), ProgressiveCountry)
+        put(ProgressiveHouse.name.lowercase(), ProgressiveHouse)
+        put(ProgressiveMetal.name.lowercase(), ProgressiveMetal)
+        put(ProgressiveRock.name.lowercase(), ProgressiveRock)
+        put(ProgressiveTrance.name.lowercase(), ProgressiveTrance)
+        put(Protopunk.name.lowercase(), Protopunk)
+        put(Psychedelic.name.lowercase(), Psychedelic)
+        put(PsychedelicPop.name.lowercase(), PsychedelicPop)
+        put(PsychedelicRock.name.lowercase(), PsychedelicRock)
+        put(PsyTrance.name.lowercase(), PsyTrance)
+        put(Punk.name.lowercase(), Punk)
+        put(PunkRock.name.lowercase(), PunkRock)
+        put(PFunk.name.lowercase(), PFunk)
+        put(Ragga.name.lowercase(), Ragga)
+        put(Rap.name.lowercase(), Rap)
+        put(Rapcore.name.lowercase(), Rapcore)
+        put(Reggae.name.lowercase(), Reggae)
+        put(Reggaeton.name.lowercase(), Reggaeton)
+        put(ReggaeFusion.name.lowercase(), ReggaeFusion)
+        put(Renaissance.name.lowercase(), Renaissance)
+        put(Rock.name.lowercase(), Rock)
+        put(Rockabilly.name.lowercase(), Rockabilly)
+        put(Rocksteady.name.lowercase(), Rocksteady)
+        put(RockAndRoll.name.lowercase(), RockAndRoll)
+        put(RootsReggae.name.lowercase(), RootsReggae)
+        put(RootsRock.name.lowercase(), RootsRock)
+        put(RAndB.name.lowercase(), RAndB)
+        put(Salsa.name.lowercase(), Salsa)
+        put(Samba.name.lowercase(), Samba)
+        put(Schlager.name.lowercase(), Schlager)
+        put(Schranz.name.lowercase(), Schranz)
+        put(Screamo.name.lowercase(), Screamo)
+        put(Shoegaze.name.lowercase(), Shoegaze)
+        put(SingerSongwriter.name.lowercase(), SingerSongwriter)
+        put(Ska.name.lowercase(), Ska)
+        put(Skacore.name.lowercase(), Skacore)
+        put(SkaPunk.name.lowercase(), SkaPunk)
+        put(Skatepunk.name.lowercase(), Skatepunk)
+        put(Skiffle.name.lowercase(), Skiffle)
+        put(Slowcore.name.lowercase(), Slowcore)
+        put(SludgeMetal.name.lowercase(), SludgeMetal)
+        put(SmoothJazz.name.lowercase(), SmoothJazz)
+        put(SoftRock.name.lowercase(), SoftRock)
+        put(Soul.name.lowercase(), Soul)
+        put(Soundtrack.name.lowercase(), Soundtrack)
+        put(SouthernRock.name.lowercase(), SouthernRock)
+        put(SpaceRock.name.lowercase(), SpaceRock)
+        put(SpeedGarage.name.lowercase(), SpeedGarage)
+        put(SpeedMetal.name.lowercase(), SpeedMetal)
+        put(Speedcore.name.lowercase(), Speedcore)
+        put(StageAndScreen.name.lowercase(), StageAndScreen)
+        put(StonerMetal.name.lowercase(), StonerMetal)
+        put(StonerRock.name.lowercase(), StonerRock)
+        put(SwampBlues.name.lowercase(), SwampBlues)
+        put(SwampPop.name.lowercase(), SwampPop)
+        put(SwampRock.name.lowercase(), SwampRock)
+        put(Swing.name.lowercase(), Swing)
+        put(SymphonicMetal.name.lowercase(), SymphonicMetal)
+        put(SymphonicRock.name.lowercase(), SymphonicRock)
+        put(Synthwave.name.lowercase(), Synthwave)
+        put(SynthPop.name.lowercase(), SynthPop)
+        put(SynthPunk.name.lowercase(), SynthPunk)
+        put(Tango.name.lowercase(), Tango)
+        put(TechnicalMetal.name.lowercase(), TechnicalMetal)
+        put(Techno.name.lowercase(), Techno)
+        put(Techstep.name.lowercase(), Techstep)
+        put(TechHouse.name.lowercase(), TechHouse)
+        put(TechTrance.name.lowercase(), TechTrance)
+        put(ThrashMetal.name.lowercase(), ThrashMetal)
+        put(Trance.name.lowercase(), Trance)
+        put(Trap.name.lowercase(), Trap)
+        put(TribalHouse.name.lowercase(), TribalHouse)
+        put(TripHop.name.lowercase(), TripHop)
+        put(TripRock.name.lowercase(), TripRock)
+        put(TropicalHouse.name.lowercase(), TropicalHouse)
+        put(Turntablism.name.lowercase(), Turntablism)
+        put(TwoStep.name.lowercase(), TwoStep)
+        put(VocalHouse.name.lowercase(), VocalHouse)
+        put(VocalJazz.name.lowercase(), VocalJazz)
+        put(VocalTrance.name.lowercase(), VocalTrance)
+        put(World.name.lowercase(), World)
+        put(WorldFusion.name.lowercase(), WorldFusion)
+    }
+}
+
+/**
+ * Joins a set of genres into a comma-separated string, sorted naturally by name.
+ *
+ * @param genres the genres to join
+ * @return a sorted, comma-separated string, or an empty string if [genres] is empty
+ */
+fun joinGenres(genres: Set<Genre>): String = genres.map(Genre::name).sorted().joinToString(", ")
+
+/**
+ * Parses a genre string into a set of [Genre] instances.
+ *
+ * Supports comma-separated genre tags (e.g. "Rock, Alternative"). Each segment is
+ * matched case-insensitively against the known genre list. Unrecognized segments
+ * are wrapped in [Genre.Custom] to preserve their original value.
+ *
+ * @param value the raw genre string from audio metadata or JSON persistence
+ * @return a set of resolved genres, or an empty set if [value] is blank
+ */
+fun parseGenre(value: String): Set<Genre> {
+    if (value.isBlank()) return emptySet()
+    return value.split(",")
+        .map { it.trim() }
+        .filter { it.isNotEmpty() }
+        .map { part -> BY_NAME[part.lowercase()] ?: Genre.Custom(part) }
+        .toSet()
+}
